jblog 평과 과제 #8
- 카테고리에 글이 남아있을 경우 삭제 실패 (해결)

+++ jblog03/src/main/java/jblog/controller/api/BlogController.java
@@ -1,6 +1,7 @@
 package jblog.controller.api;
 
 import jblog.service.CategoryService;
+import jblog.service.PostService;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -9,13 +10,19 @@ import org.springframework.web.bind.annotation.*;
 //@Authorization(role = "ADMIN")
 public class BlogController {
     private final CategoryService categoryService;
+    private final PostService postService;
 
-    public BlogController(CategoryService categoryService) {
+    public BlogController(CategoryService categoryService, PostService postService) {
         this.categoryService = categoryService;
+        this.postService = postService;
     }
 
     @GetMapping({"/category/delete/{id}"})
     public ResponseEntity<String> adminCategoryDelete(@PathVariable("id") int categoryId) {
+        if (!postService.existByCategoryId(categoryId)) {
+            return ResponseEntity.status(510).body("게시글이 있어 삭제할 수 없습니다.");
+        }
+
         if (!categoryService.deleteCategory(categoryId)){
             return ResponseEntity.status(501).body("기본 카테고리는 삭제할 수 없습니다.");
         }
